requirements.txt
django
djangorestframework
drf-yasg
psycopg2-binary
django-filter
djangorestframework-jwt
djangorestframework-simplejwt
httpx
django-json-widget
django-admin-rangefilter
ruff
django_crontab
